Code cleanup + fixups

+++ lib/comms/src/comms.h
@@ -72,7 +72,7 @@ static const uint32_t crc_table[256] = {
   0xbcb4666d, 0xb8757bda, 0xb5365d03, 0xb1f740b4
 };
 
-void comms_send(uint8_t data[], uint16_t data_length);
+void comms_send(uint8_t data[], uint8_t data_length);
 int comms_recv(char *data[]);
 void comms_imu(vec3_t mag, vec3_t accel, vec3_t gyro, float hoz);
 void comms_env(float temp, float humidity, float pressure);
